Remove tests/common.py
